urldecode: improve handling of invalid % sequences
These sequences should be simply passed through, here's my reasoning.

An invalid sequence shows the encoder was bad, and ideally that'd be
conspicuous so it'd be noticed and fixed. But that's well outside the
control of urldecode's implementation, since it's a fairly low-level
routine on a little device that is likely to be tucked away somewhere,
silently doing its job.

Probably in most cases, such input is so bad that there's nothing to be
done except for the whole operation (e.g. HTTP request) to fail. To that
end, urldecode could signal failure in-band by generating an empty
string, or out-of-band by returning an int, etc. - clean and reasonable.
However, that creates a burden at each callsite to check and handle the
failure, and does not leave behind much in the way of diagnostic
information.

It's also not a "best effort." When these cases occur, there's a
significant chance that there's a real user somewhere behind that bad
encoder, whose intent is reflected in a string like "Alcohol 120%" or
"file%G0name". In the general case, we can't recover such intentions, so
trying will only cause sadness. But we can be pretty sure the intention
wasn't to perform an operation on a file named "Alcohol 120" or
"filename" or "file%name" on the SD card. Meanwhile, if the invalid
sequence is visible in places like logs, it's more obvious. So passing
through the invalid sequence is a nicer default, here.

Ideally, the routine would also return a value reflecting the validity
of the input, so that the caller could check this and handle it as they
desire. That's a topic for another day.

This commit takes a small step from the existing code. It only checks
the return value from sscanf. If sscanf reached end of string
prematurely (strings ending in "%") or if it encountered a format
mismatch (first character not a hex digit) then it returns a
non-positive value, and urldecode just writes the "%". This should
handle cases like "%%" and "%G0".

The residual cases are ones where sscanf gets at least one valid hex
digit, and then something else happens, like end of string or a
character which isn't a hex digit. Examples: "%A", "%1G". In these
cases, sscanf writes a byte based on the hex digit it did get
(ASCII 10 for "A", ASCII 1 for "1"). sscanf sees nothing wrong with this
and returns a positive value, so checking its return value doesn't help.
This commit doesn't fix that.