urldecode: fix handling of invalid % sequences
This commit passes through invalid codes in cases starting with a valid
hex digit, like "%A" and "%1G", which sscanf will uncritically accept.
It still uses the same call to sscanf as before, it just puts that
inside a conditional ensuring there are two hex digits to be read. If
there aren't, it just writes the "%" and lets the loop move on, copying
any further characters.

This adds a dependency: <ctype.h> for isxdigit. Maybe a C++ programmer
would prefer <cctype>.