An attempt to load all the filenames in one go
This try to cache all the filenames into memory before the filenames are
requested. That way the serving of filenames for the load images page
doesn't have to got through the I2C bus to fetch image filenames on the
ZuluIDE and the serve each filename as a separate request.

In its current state there seem to be a memory leak as it only works
with a couple images on the SD card before "UNABLE TO GET A FREE BUFFER"
is spammed to the USB logging on the Pico W webserver shield.