Make the I2C communication run on the second core
Because all the data between the I2C side and the LWIP HTTPd side goes
through thread and irq safe message queuing, it makes sense to
separate the two on separate cores. The http server seems much
more responsive now.

There are some regressions that still need to be fixed, for example
the when the UI is first loaded, the current image is blank instead
of showing what is loaded on the ZuluIDE.

Also the UI should tell the user when the filenames are still being
generated, or when the is an overflow of the filenames in the cache
and the UI is falling back to the single file queries.