Fix urldecode Fixes urldecode to work correctly when run on the RP2040 itself.
Previously, the tests I added in this commit passed with the sscanf
urldecode when run on my desktop. However, it was broken when run on the
RP2040. Examples from the serial console:

    Setting image to: Game %USAS Europer
    Setting image to: Rebel Moon Rising %USAS %En+Fr8De%EsFIt2
    Setting image to: Windows 98 Second Edition %OEME.iso

After this change, urldecode works as expected on both my desktop and
the RP2040.

    Setting image to: Game (USA, Europe)
    Setting image to: Rebel Moon Rising (USA) (En,Fr,De,Es,It)
    Setting image to: Windows 98 Second Edition (OEM).iso

I'm not sure why exactly sscanf did not work correctly, but some vague
searches seem to indicate that the `hh` modifier of `%2hhx` is not
supported by the pi pico sdk.

The test structure was copied from the cueparser library.