Use the MCU LED for boot indication
The Audio + RM2 Shield has a LED connected directly to the MCU.
This LED is turned on as the very start of main after it detects the
pull down resistor on GPIO 5 which indicates this board the the
Audio + RM2 and not the Audio + Pico Shield.

The MCU LED then blinks in sync with the RM2 LED three times if the
board is not connected the a ZuluIDE, otherwise it will turn off
once the ZuluIDE starts communicating with the Audio + RM2 Shield.

I